fix(unit-tests): Use IoHelper::setRights when changing permissions on Windows (Hervé's CR comment)
