+++ test/libcommonserver/io/testchecksetgetrights.cpp
@@ -17,6 +17,7 @@
  */
 
 #include "testio.h"
+#include "test_utility/testhelpers.h"
 
 #include <filesystem>
 
@@ -24,19 +25,7 @@ using namespace CppUnit;
 
 namespace KDC {
 
-struct RightsSet {
-        RightsSet(int rights) :
-            read(rights & 4),
-            write(rights & 2),
-            execute(rights & 1) {};
-        RightsSet(bool read, bool write, bool execute) :
-            read(read),
-            write(write),
-            execute(execute) {};
-        bool read;
-        bool write;
-        bool execute;
-};
+using namespace testhelpers;
 
 void TestIo::testCheckSetAndGetRights() {
     // Test if the rights are correctly set and get on a directory
