+++ test/test_utility/testhelpers.h
@@ -84,6 +84,20 @@ struct TestVariables {
         }
 };
 
+struct RightsSet {
+        RightsSet(int rights) :
+            read(rights & 4),
+            write(rights & 2),
+            execute(rights & 1){};
+        RightsSet(bool read, bool write, bool execute) :
+            read(read),
+            write(write),
+            execute(execute){};
+        bool read;
+        bool write;
+        bool execute;
+};
+
 void generateOrEditTestFile(const SyncPath &path);
 /**
  * @brief Generate test files.
